Move controller definition and add exception
