+++ src/Core/Framework/Resources/config/routes.xml
@@ -16,5 +16,6 @@
     <import resource="../../Script/Api/*Route.php" type="attribute" />
     <import resource="../../Rule/Api/*Controller.php" type="attribute" />
     <import resource="../../Gateway/**/*Route.php" type="attribute" />
+    <import resource="../../Sso/Controller/**/*Controller.php" type="attribute" />
 
 </routes>
